restructure panel and panel list components
I tried to re-structure the panel and panel-list components so they didn't rely on Ember.Evented anymore.
Biggest changes:
- Use the block content feature. This allows to make the component a bit more generic (premature optimization, but I also think it makes usage cleaner)
- Simplified data flow and management. Panel List now figures out what needs to expand or filter out, or not. Panel becomes a template-only component
- Filtering how hides HTML elements with CSS. I realized that doing it with template conditionals meant that we would re-render all the components every time the query changed. This way we avoid doing unnecessary work.
