+++ app/components/banlist/tabs.gjs
@@ -1,11 +1,9 @@
 import { LinkTo } from '@ember/routing';
-import { fn, hash } from '@ember/helper';
-
-import LinkToCard from '../card/link-to';
-import PanelList from '../panel-list';
-import Panel from '../panel';
-import formatDate from '../../helpers/format-date';
+import { hash } from '@ember/helper';
 import { eq } from '../../utils/template-operators';
+import Accordion from '../ui/accordion';
+import CardLinkTo from '../card/link-to';
+import formatDate from '../../helpers/format-date';
 
 <template>
   <div class="card flex-grow-1" id="nav-main-content">
@@ -26,51 +24,56 @@ import { eq } from '../../utils/template-operators';
       {{#each @formats as |format|}}
         {{#if (eq @selectedFormat format.id)}}
           <div id={{format.id}}>
-            {{#if format.restrictions}}
-              <PanelList
-                @showMassToggle="true"
-                @showSearch="true"
-                @items={{format.restrictions}}
-                as |restriction isOpen actions|
-              >
-                <Panel
-                  @title="{{restriction.name}}"
-                  @subtitle="{{restriction.size}} cards. Start Date: {{formatDate restriction.dateStart}}."
-                  @isOpen={{isOpen}}
-                  @showPanel={{fn actions.showPanel restriction.name}}
-                  @hidePanel={{fn actions.hidePanel restriction.name}}
-                >
-                  <div class="row">
-                    <div class="col-6">
-                      <h3>Corp Cards</h3>
-                      <ul>
-                        <li>
-                          <strong>Banned</strong>
-                          <ul>
-                            {{#each restriction.verdicts.banned as |banned|}}
-                              <li>
-                                <LinkToCard
-                                  @id={{banned}}
-                                  class="text-truncate"
-                                >
-                                  {{banned}}
-                                </LinkToCard>
-                              </li>
-                            {{/each}}
-                          </ul>
-                        </li>
-                      </ul>
+            <Accordion
+              @showMassToggle={{true}}
+              @showSearch={{true}}
+              @expandFirstItem={{true}}
+              @items={{format.restrictions}}
+            >
+              <:default as |Panel item|>
+                <Panel>
+                  <:title>
+                    {{item.name}}
+                  </:title>
+                  <:subtitle>
+                    {{item.size}}
+                    cards. Start Date:
+                    {{(formatDate item.dateStart)}}.
+                  </:subtitle>
+                  <:body>
+                    <div class="row">
+                      <div class="col-6">
+                        <h3>Corp Cards</h3>
+                        <ul>
+                          <li>
+                            <strong>Banned</strong>
+                            <ul>
+                              {{#each item.verdicts.banned as |banned|}}
+                                <li>
+                                  <CardLinkTo
+                                    @id={{banned}}
+                                    class="text-truncate"
+                                  >
+                                    {{banned}}
+                                  </CardLinkTo>
+                                </li>
+                              {{/each}}
+                            </ul>
+                          </li>
+                        </ul>
+                      </div>
+                      <div class="col-6">
+                        <h3>Runner Cards</h3>
+                      </div>
                     </div>
-                    <div class="col-6">
-                      <h3>Runner Cards</h3>
-                    </div>
-                  </div>
+                  </:body>
                 </Panel>
-              </PanelList>
-            {{else}}
-              No cards are currently banned in
-              {{format.name}}. Have a blast!
-            {{/if}}
+              </:default>
+              <:empty>
+                No cards are currently banned in
+                {{format.name}}. Have a blast!
+              </:empty>
+            </Accordion>
           </div>
         {{/if}}
       {{/each}}
