--- app/components/panel.hbs
@@ -1,24 +1,0 @@
-{{#unless this.filteredOut}}
-  <div class="panel panel-default {{if @isOpen 'panel-open'}}">
-    <div class="panel-heading">
-      <div class="panel-heading-text">
-        <h4 class="panel-heading-text-title">{{@title}}</h4>
-        {{#if @subtitle}}
-          <h6 class="panel-heading-text-subtitle">
-            {{@subtitle}}
-          </h6>
-        {{/if}}
-      </div>
-      <button
-        class="btn btn-secondary"
-        type="button"
-        {{on "click" this.togglePanel}}
-      >{{if @isOpen "Hide" "Show"}}</button>
-    </div>
-    {{#if @isOpen}}
-      <div class="panel-body">
-        {{yield}}
-      </div>
-    {{/if}}
-  </div>
-{{/unless}}
